requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.0.1
Django==6.0
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
django-js-asset==3.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
whitenoise==6.8.2
